+++ firefoxThemes/popup/popup.js
@@ -19,13 +19,15 @@ const backgroundTypes = {
 /*Set color of popup menu based on theme*/
 function setCss(chosenTheme) {
   if (!chosenTheme) return
-  const {colors} = chosenTheme.definition
+  const {colors, information} = chosenTheme.definition;
+  const switchShadowColor = information.dark ? 'white' : 'black';
+  const dokiShadowProps = information.dark ? '11px #fff' : '11px #000';
+
   const styles = `
 .popup-header {
   background-color: ${colors.headerColor};
   color: ${colors.infoForeground};
 }
-
 
 * {
  color: ${colors.lineNumberColor};
@@ -59,16 +61,27 @@ input:checked + .slider {
 input:focus + .slider {
   box-shadow: 0 0 1px ${colors.accentColor}44;
 }
+
+input:not(:checked):not(:disabled) + .slider
+{
+  filter: drop-shadow(0 0 3px ${switchShadowColor});
+}
+
 label[for="backgroundType"],
 label[for="darkMode"],
 label[for="hideSearch"]
 {
   color:${colors.infoForeground};
 }
-        `
-  const styleSheet = document.createElement("style");
-  styleSheet.innerText = styles;
-  document.head.appendChild(styleSheet)
+
+img:hover{
+  filter: hue-rotate(90deg) drop-shadow(0px 0px ${dokiShadowProps});
+}
+`;
+  const styleTag = document.createElement("style");
+  const styleText = document.createTextNode(styles);
+  styleTag.append(styleText);
+  document.head.append(styleTag);
 }
 
 const setBackground = async () => {
@@ -89,11 +102,26 @@ const setHideWidget = async () => {
   browser.runtime.sendMessage({currentThemeId});
 }
 
-const setDarkMode = async () => {
+/*Stores info to set dark mode switch accordingly*/
+async function setDarkMode(shouldDisable) {
   await browser.storage.local.set({
-    darkMode: darkModeSwitch.checked
+    darkMode: {
+      isDarkNow: darkModeSwitch.checked,
+      shouldDisable
+    }
   });
+}
 
+/*Stores whether or not the current theme has a secondary theme*/
+function setHasSecondary(currentTheme) {
+  const hasSecondaryMode = !!currentTheme.backgrounds.secondary;
+  backgroundSwitch.disabled = !hasSecondaryMode;
+  browser.storage.local.set({hasSecondaryMode});
+}
+
+/*Set the theme to the opposite type. (Light vs Dark)
+* Ex: If current theme is light, go to dark and vice versa*/
+const setOpposingTheme = async () => {
   const {currentThemeId, waifuThemes} = await browser.storage.local.get(["currentThemeId", "waifuThemes"]);
 
   const currentTheme = waifuThemes.themes[currentThemeId];
@@ -103,10 +131,10 @@ const setDarkMode = async () => {
       .find(dokiTheme =>
         (dokiTheme.displayName === currentTheme.displayName ||
           dokiTheme.name === currentTheme.name) &&
-        dokiTheme.id !== currentThemeId) || currentTheme
+        dokiTheme.id !== currentThemeId) || currentTheme;
     const newThemeId = newTheme.id;
     setCss(newTheme);
-    await browser.storage.local.set({currentThemeId: newThemeId})
+    await browser.storage.local.set({currentThemeId: newThemeId});
     browser.runtime.sendMessage({currentThemeId: newThemeId});
   }
 }
@@ -129,15 +157,18 @@ function setTheme(e) {
             dokiTheme.name === chosenThemeName
           ));
 
-        const isDark = storage.darkMode !== undefined && storage.darkMode;
+        const isDark = (storage.darkMode) && storage.darkMode.isDarkNow;
         const theme = themes.find(dokiTheme =>
           dokiTheme.dark === isDark);
 
-        darkModeSwitch.disabled = themes.length < 2;
+        const disableDarkSwitch = themes.length < 2;
+        darkModeSwitch.disabled = disableDarkSwitch;
+        setDarkMode(disableDarkSwitch);
+
 
         if (chosenThemeName !== 'random') {
           const usableTheme = theme || themes[0];
-          darkModeSwitch.checked = usableTheme.dark;
+          darkModeSwitch.checked = !!usableTheme.dark;
 
           if (usableTheme) {
             setCss(usableTheme);
@@ -145,6 +176,7 @@ function setTheme(e) {
           }
         }
       }
+      setHasSecondary(storage.waifuThemes.themes[chosenThemeId]);
       browser.runtime.sendMessage({currentThemeId: chosenThemeId || 'mixed', mixState: currentMix});
     });
 }
@@ -160,10 +192,26 @@ function getRandomTheme(themes) {
 function getRandomNumber(min, max) {
   return Math.floor(Math.random() * (max - min)) + min;
 }
+
 /*Navigate to options page*/
-function optionsPage(){
+function optionsPage() {
   browser.runtime.openOptionsPage();
 }
+
+/*Initializes checkbox switches*/
+function prepareSwitches(storage) {
+  backgroundSwitch.checked = !!storage.backgroundType;
+  backgroundSwitch.disabled = !(!!storage.hasSecondaryMode);
+  if (storage.darkMode !== undefined) {
+    darkModeSwitch.checked = !!storage.darkMode.isDarkNow
+    darkModeSwitch.disabled = !!storage.darkMode.shouldDisable;
+  } else {
+    darkModeSwitch.checked = false;
+    darkModeSwitch.disabled = true;
+  }
+  showSearchSwitch.checked = !!storage.showWidget;
+}
+
 /*Setup Waifu Choices for the popup menu
 * Also categorizes each theme based on their type (dark/light)*/
 function initChoice() {
@@ -173,12 +221,11 @@ function initChoice() {
     "mixedTabs",
     "backgroundType",
     'showWidget',
-    "darkMode"
+    "darkMode",
+    "hasSecondaryMode"
   ])
     .then((storage) => {
-      backgroundSwitch.checked = !!storage.backgroundType;
-      darkModeSwitch.checked = storage.darkMode;
-      showSearchSwitch.checked = storage.showWidget === undefined || storage.showWidget;
+      prepareSwitches(storage);
       const themesGroupedByName = Object.values(storage.waifuThemes.themes)
         .reduce((accum, dokiTheme) => {
           const displayName = dokiTheme.displayName;
@@ -207,7 +254,7 @@ function initChoice() {
       themes.forEach(themeName => {
         const themeOption = document.createElement("option");
         themeOption.setAttribute("value", themeName);
-        themeOption.id = themeName
+        themeOption.id = themeName;
         const txtNode = document.createTextNode(themeName);
         themeOption.append(txtNode);
         waifuGroup.append(themeOption);
@@ -234,5 +281,5 @@ initChoice();
 selectTag.addEventListener("change", setTheme, true);
 backgroundSwitch.addEventListener("change", setBackground, true);
 showSearchSwitch.addEventListener("change", setHideWidget, true);
-darkModeSwitch.addEventListener("change", setDarkMode, true);
-dokiHeart.addEventListener('click',optionsPage,true);
+darkModeSwitch.addEventListener("change", setOpposingTheme, true);
+dokiHeart.addEventListener('click', optionsPage, true);